+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Src
-- 🎉 added ayu, ayu_mirage, synthwave, material and github themes!
+- 🎉 added ayu, ayu_mirage, synthwave, material, rose_pine, kanagawa, vscode, everforest and github themes!
 - 📈 markdown viewer now uses the theme colors and not terminal colors
 - 📈 improved the markdown viewer -- less clutter